Adapt AutoEnchant module for existing client
